Refactring component postView
